Update Readme and Adapt some Utilities
